requirements.txt
aiohttp>=3.10.0
python-dotenv>=1.0.0
typing-extensions>=4.1.0
aiohappyeyeballs>=2.3.0
multidict>=6.0.0
yarl>=1.9.0
aiosignal>=1.3.0
async-timeout>=4.0.0
frozenlist>=1.4.0
